+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - You can now collect payment details before creating a `PaymentIntent` or `SetupIntent`. See [our docs](https://stripe.com/docs/payments/accept-a-payment-deferred?platform=react-native) for more info. This integration also allows you to [confirm the Intent on the server]
 (https://stripe.com/docs/payments/finalize-payments-on-the-server?platform=react-native).
 
+## Fixes
+
+- Fixes `handleURLCallback` to only take action on iOS, no-op on Android. [#1423](https://github.com/stripe/stripe-react-native/pull/1423)
+
 ## 0.28.0 - 2023-06-16
 
 ## Features